+++ src/ethereum/ethereumTypes.ts
@@ -137,6 +137,7 @@ export type EthereumBaseMultiplier = ReturnType<
 export type KeysOfEthereumBaseMultiplier = keyof EthereumBaseMultiplier
 
 export const asEthereumFee = asObject({
+  baseFee: asOptional(asString),
   baseFeeMultiplier: asOptional(asEthereumBaseFeeMultiplier),
   gasLimit: asOptional(asEthereumFeesGasLimit),
   gasPrice: asOptional(asEthereumFeesGasPrice),
@@ -284,6 +285,7 @@ export interface EthereumTxOtherParams {
   gas: string
   gasPrice: string
   gasUsed: string
+  minerTip?: string
   tokenRecipientAddress?: string
   nonceUsed?: string
   replacedTxid?: string
@@ -296,6 +298,7 @@ export const asEthereumTxOtherParams = asObject<EthereumTxOtherParams>({
   gas: asString,
   gasPrice: asString,
   gasUsed: asString,
+  minerTip: asOptional(asString),
   tokenRecipientAddress: asOptional(asString),
   nonceUsed: asOptional(asString),
   replacedTxid: asOptional(asString),