+++ src/ethereum/info/binancesmartchainInfo.ts
@@ -27,6 +27,7 @@ const builtinTokens: EdgeTokenMap = {
 
 const defaultNetworkFees: EthereumFees = {
   default: {
+    baseFee: undefined,
     baseFeeMultiplier: undefined,
     gasLimit: {
       regularTransaction: '21000',