+++ src/ethereum/info/rinkebyInfo.ts
@@ -10,6 +10,7 @@ const builtinTokens: EdgeTokenMap = {}
 
 const defaultNetworkFees: EthereumFees = {
   default: {
+    baseFee: undefined,
     baseFeeMultiplier: {
       lowFee: '1',
       standardFeeLow: '1.25',
@@ -33,6 +34,7 @@ const defaultNetworkFees: EthereumFees = {
     minPriorityFee: '2000000000'
   },
   '1983987abc9837fbabc0982347ad828': {
+    baseFee: undefined,
     baseFeeMultiplier: undefined,
     // @ts-expect-error
     gasLimit: {
@@ -51,6 +53,7 @@ const defaultNetworkFees: EthereumFees = {
     minPriorityFee: undefined
   },
   '2983987abc9837fbabc0982347ad828': {
+    baseFee: undefined,
     baseFeeMultiplier: undefined,
     // @ts-expect-error
     gasLimit: {