+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.17.2 (2023-12-11)
+
 - fixed: Replace 2 gwei priority fee minimum non-negative minimum
 - changed: Query and cache network base-fee for EIP1559 currencies